Stand-alone Julia code for synchrotron spectrum and some documentation (#2)
* Synchrotron calculation

Updated synchrotron calculation to plot the emissivity as a function of photon frequency. Note that the code is not complete, but I hope this clarifies things a bit.

* Update .gitignore

Not sure if we should include the Project.toml file or not - I have included it for now!

* Synchrotron Julia code and documentation

Started some stand-alone Julia code with documentation.