Automatic documentation (#8)
* Synchrotron calculation

Updated synchrotron calculation to plot the emissivity as a function of photon frequency. Note that the code is not complete, but I hope this clarifies things a bit.

* Update .gitignore

Not sure if we should include the Project.toml file or not - I have included it for now!

* Synchrotron Julia code and documentation

Started some stand-alone Julia code with documentation.

* Fixed critical frequency and added plot test

Fixed typo in critical frequency value.

Added plotTest.jl which simply calls syncPlot()

* Very minor updates

Very minor updates

* Automatic documentation

An attempt to get GitHub to automatically create the documentation on the gh-pages branch.

* Update index.md

Added S_syn documentation

* Update Documentation.yml

"master" branch is called "main" - this might have prevented the script from running

* Updated documentation

Also noted that the S_syn function doesn't take any parameters

* Update make.jl

Change repo to the astro-group-bristol repo.